+++ jest.config.js
@@ -1,3 +1,8 @@
+// Force use of modules in Babel when running tests
+// with an external tool like vscode-jest.
+process.env.VUE_CLI_BABEL_TARGET_NODE = true
+process.env.VUE_CLI_BABEL_TRANSPILE_MODULES = true
+
 module.exports = {
   preset: '@vue/cli-plugin-unit-jest/presets/typescript-and-babel',
   transformIgnorePatterns: [
@@ -7,5 +12,6 @@ module.exports = {
     '^@/(.*)$': '<rootDir>/src/$1',
     '^.+.(css|styl|less|sass|scss|png|jpg|ttf|woff|woff2)$': 'jest-transform-stub'
   },
+  setupFiles: ['<rootDir>/tests/unit/setup.js'],
   collectCoverage: true
 }