+++ src/store/actions.ts
@@ -6,13 +6,13 @@ import Assessment from '@/types/Assessment'
 import { Section } from '@/types/Section.ts'
 import { Cart } from '@/types/Cart'
 import ApplicationState from '@/types/ApplicationState'
-import router from '@/router'
 import Getters from '@/types/Getters'
 import { buildFormData, generateOrderNumber } from '@/services/orderService.ts'
 import FormField from '@/types/FormField'
 import { OrderState } from '@/types/Order'
 import moment from 'moment'
 import { TreeParent } from '@/types/Tree'
+import axios from 'axios'
 
 const buildPostOptions = (formData: any, formFields: FormField[]) => {
   return {
@@ -55,6 +55,9 @@ const createOrder = async (formData: any, formFields:FormField[]) => {
 }
 
 const updateOrder = async (formData: any, formFields:FormField[]) => {
+  if (formData.applicationForm && (typeof formData.applicationForm.filename === 'string')) {
+    formData.applicationForm = formData.applicationForm.filename
+  }
   const options = buildPostOptions(formData, formFields)
 
   return api.post(`/api/v1/lifelines_order/${formData.orderNumber}?_method=PUT`, options, true).then(() => {
@@ -71,6 +74,7 @@ export default {
   deleteOrder: tryAction(async ({ dispatch, commit }: any, orderId: string) => {
     commit('setOrders', null)
     await api.delete_(`/api/v2/lifelines_order/${orderId}`)
+    commit('setToast', { type: 'success', message: `Deleted order with order number ${orderId}` })
     dispatch('loadOrders')
   }),
   loadSections: tryAction(async ({ commit, state } : any) => {
@@ -150,8 +154,7 @@ export default {
     commit('updateVariables', variableMap)
   }),
   loadGridVariables: tryAction(async ({ state, commit, getters } : { state: ApplicationState, commit: any, getters: Getters}) => {
-    state.isGridLoading = true
-    commit('updateGridVariables', [])
+    commit('updateGridVariables', null)
     const subsectionId = state.treeSelected
     const searchTermQuery = getters.searchTermQuery
     let q = `subsection_id==${subsectionId}`
@@ -171,7 +174,6 @@ export default {
             }))
         })))
     }
-    state.isGridLoading = false
   }),
   loadParticipantCount: tryAction(async ({ commit, getters }: any) => {
     commit('updateParticipantCount', null)
@@ -186,7 +188,7 @@ export default {
     }
   }),
   loadGridData: tryAction(async ({ commit, getters }: any) => {
-    commit('updateVariantCounts', [])
+    commit('updateVariantCounts', null)
     let url = '/api/v2/lifelines_who_when?aggs=x==variant_id'
     const rsql = getters.rsql
     if (rsql) {
@@ -212,7 +214,7 @@ export default {
     if (state.order.orderNumber) {
       await updateOrder(formData, formFields)
       commit('setToast', { type: 'success', message: 'Saved order with order number ' + state.order.orderNumber })
-      router.push({ name: 'load', params: { orderNumber: state.order.orderNumber } })
+      return state.order.orderNumber
     } else {
       const creationDateField = { id: 'creationDate', type: 'date' }
       const orderNumber = await createOrder(formData, [ ...formFields, creationDateField ]).catch(() => {
@@ -221,10 +223,10 @@ export default {
       const newOrderResponse = await api.get(`/api/v2/lifelines_order/${orderNumber}`)
       commit('restoreOrderState', newOrderResponse)
       commit('setToast', { type: 'success', message: 'Saved order with order number ' + orderNumber })
-      router.push({ name: 'load', params: { orderNumber: orderNumber } })
+      return orderNumber
     }
   }),
-  submit: tryAction(async ({ state, commit }: {state: ApplicationState, commit: any}) => {
+  submit: tryAction(async ({ state, commit, dispatch }: {state: ApplicationState, commit: any, dispatch: any}) => {
     const formFields = [...state.orderFormFields, { id: 'contents', type: 'text' }]
     const now = moment().toISOString()
     const formData = {
@@ -247,8 +249,9 @@ export default {
     }
     const newOrderResponse = await api.get(`/api/v2/lifelines_order/${orderNumber}`)
     commit('restoreOrderState', newOrderResponse)
+    dispatch('givePermissionToOrder')
+    dispatch('sendSubmissionTrigger')
     commit('setToast', { type: 'success', message: 'Submitted order with order number ' + orderNumber })
-    router.push({ name: 'orders' })
   }),
   load: tryAction(async ({ state, commit }: {state: ApplicationState, commit: any}, orderNumber: string) => {
     const response = await api.get(`/api/v2/lifelines_order/${orderNumber}`)
@@ -258,5 +261,30 @@ export default {
     commit('updateFacetFilter', facetFilter)
     commit('updateGridSelection', gridSelection)
     commit('setToast', { type: 'success', message: 'Loaded order with orderNumber ' + orderNumber })
-  })
+  }),
+  givePermissionToOrder: tryAction(async ({ state, commit }: {state: ApplicationState, commit: any}, orderNumber: string) => {
+    const data = {
+      objects: [{
+        objectId: state.order.orderNumber,
+        permissions: [
+          {
+            role: 'LIFELINES_MANAGER',
+            permission: 'WRITE'
+          }
+        ]
+      }]
+    }
+    const options = {
+      body: JSON.stringify(data)
+    }
+
+    return api.post('/api/permissions/entity-lifelines_order', options)
+  }),
+  sendSubmissionTrigger: async () => {
+    return axios.post('/edge-server/trigger?type=submit').catch((err: any) => {
+      /* eslint no-console: ["error", { allow: ["warn", "error"] }] */
+      console.warn('Send submit trigger failed')
+      console.warn(err)
+    })
+  }
 }