+++ src/store/state.ts
@@ -1,8 +1,6 @@
-import ApplicationState from '@/types/ApplicationState'
+import { AppState } from '@/types/ApplicationState'
 
-const state: ApplicationState = {
-  isContextLoaded: false,
-  isSignedIn: true, // set true during development
+const state: AppState = {
   toast: null,
   order: {
     orderNumber: null,
@@ -50,7 +48,7 @@ const state: ApplicationState = {
     }
   ],
   variables: {},
-  gridVariables: [],
+  gridVariables: null,
   genderOptions: [
     { value: '1', text: 'Male' },
     { value: '2', text: 'Female' }
@@ -86,13 +84,12 @@ const state: ApplicationState = {
   treeOpenSection: -1,
   treeOpenPageSection: -1,
   gridSelection: {},
-  variantCounts: [],
+  variantCounts: null,
   participantCount: null,
   assessments: {},
   searchTerm: null,
   filteredSubsections: null, // contains the IDs of subsections that either match the search term or contain variables that match the search term
   filteredSections: null, // contains the IDs of sections that match the search term
-  isGridLoading: false,
   orders: null
 }