+++ tests/unit/store/actions.spec.ts
@@ -1,11 +1,11 @@
 import actions from '@/store/actions'
-import router from '@/router'
 import { Cart } from '@/types/Cart'
 import emptyState from '@/store/state'
 import orders from '../fixtures/orders'
 
 // @ts-ignore
 import { post } from '@molgenis/molgenis-api-client'
+import axios from 'axios'
 import ApplicationState from '@/types/ApplicationState'
 import { OrderState } from '@/types/Order'
 import * as orderService from '@/services/orderService'
@@ -51,6 +51,12 @@ const mockResponses: {[key:string]: Object} = {
     items: [
       { id: 1, name: 'sub_section1' },
       { id: 2, name: 'sub_section2' }
+    ]
+  },
+  '/api/v2/lifelines_tree?num=10000': {
+    items: [
+      { id: 1, section_id: { id: 1 }, subsection_id: { id: 1 } },
+      { id: 2, section_id: { id: 1 }, subsection_id: { id: 2 } }
     ]
   },
   '/api/v2/lifelines_assessment': {
@@ -191,10 +197,6 @@ jest.mock('@molgenis/molgenis-api-client', () => {
   }
 })
 
-jest.mock('@/router', () => ({
-  push: jest.fn()
-}))
-
 describe('actions', () => {
   describe('loadOrders', () => {
     it('loads the orders and commits them', async (done) => {
@@ -311,6 +313,21 @@ describe('actions', () => {
     })
   })
 
+  describe('loadSectionTree', () => {
+    it('fetches the sections in tree form', async (done) => {
+      const commit = jest.fn()
+      await actions.loadSectionTree({ state: { treeStructure: [] }, commit })
+      expect(commit).toHaveBeenCalledWith('updateSectionTree', [{ 'key': '1', 'list': [1, 2] }])
+      done()
+    })
+    it('wont fetch if data is loaded already', async (done) => {
+      const commit = jest.fn()
+      await actions.loadSectionTree({ state: { treeStructure: [0, 1] }, commit })
+      expect(commit).toHaveBeenCalledTimes(0)
+      done()
+    })
+  })
+
   describe('loadAssessments', () => {
     it('loads the assessments and commits them', async (done) => {
       const commit = jest.fn()
@@ -331,7 +348,7 @@ describe('actions', () => {
         getters: { searchTermQuery: null },
         commit
       })
-      expect(commit).toHaveBeenCalledWith('updateGridVariables', [])
+      expect(commit).toHaveBeenCalledWith('updateGridVariables', null)
       await action
       const variant = { 'assessmentId': 1, 'assessment_id': 1, 'id': 197 }
       expect(commit).toHaveBeenCalledWith('updateGridVariables', [
@@ -349,7 +366,7 @@ describe('actions', () => {
         getters: { searchTermQuery: '*=q=cream' },
         commit
       })
-      expect(commit).toHaveBeenCalledWith('updateGridVariables', [])
+      expect(commit).toHaveBeenCalledWith('updateGridVariables', null)
       await action
       const variant = { 'assessmentId': 1, 'assessment_id': 1, 'id': 197 }
       expect(commit).toHaveBeenCalledWith('updateGridVariables', [
@@ -363,7 +380,7 @@ describe('actions', () => {
       const state = { treeSelected: 4 }
       const getters = { searchTermQuery: null }
       const action = actions.loadGridVariables({ state, commit, getters })
-      expect(commit).toHaveBeenCalledWith('updateGridVariables', [])
+      expect(commit).toHaveBeenCalledWith('updateGridVariables', null)
       state.treeSelected = 6
       await action
       expect(commit).toHaveBeenCalledTimes(1)
@@ -375,7 +392,7 @@ describe('actions', () => {
       const state = { treeSelected: 4 }
       const getters: any = { searchTermQuery: null }
       const action = actions.loadGridVariables({ state, commit, getters })
-      expect(commit).toHaveBeenCalledWith('updateGridVariables', [])
+      expect(commit).toHaveBeenCalledWith('updateGridVariables', null)
       getters.searchTermQuery = '*=q=test'
       await action
       expect(commit).toHaveBeenCalledTimes(1)
@@ -431,7 +448,7 @@ describe('actions', () => {
     it('loads new variant counts if rsql is empty', async (done) => {
       const commit = jest.fn()
       const response = actions.loadGridData({ commit, getters: { rsql: '' } })
-      expect(commit).toHaveBeenCalledWith('updateVariantCounts', [])
+      expect(commit).toHaveBeenCalledWith('updateVariantCounts', null)
       await response
       expect(commit).toHaveBeenCalledWith('updateVariantCounts', [
         { 'count': 12340, 'variantId': 1 },
@@ -443,7 +460,7 @@ describe('actions', () => {
     it('loads new variant counts if rsql is nonempty', async (done) => {
       const commit = jest.fn()
       await actions.loadGridData({ commit, getters: { rsql: 'll_nr.yob=le=1970' } })
-      expect(commit).toHaveBeenCalledWith('updateVariantCounts', [])
+      expect(commit).toHaveBeenCalledWith('updateVariantCounts', null)
       expect(commit).toHaveBeenCalledWith('updateVariantCounts', [
         { 'count': 1234, 'variantId': 1 },
         { 'count': 5678, 'variantId': 10 }
@@ -455,7 +472,7 @@ describe('actions', () => {
       const commit = jest.fn()
       const getters = { rsql: 'll_nr.yob=le=1970' }
       const action = actions.loadGridData({ commit, getters })
-      expect(commit).toHaveBeenCalledWith('updateVariantCounts', [])
+      expect(commit).toHaveBeenCalledWith('updateVariantCounts', null)
       getters.rsql = ''
       await action
       expect(commit).toHaveBeenCalledTimes(1)
@@ -497,7 +514,8 @@ describe('actions', () => {
           }
         }
         post.mockResolvedValue('success')
-        await actions.save({ state, commit })
+        const response = await actions.save({ state, commit })
+        expect(response).toBe('12345')
         expect(post).toHaveBeenCalledWith('/api/v1/lifelines_order/12345?_method=PUT', expect.anything(), true)
         expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Saved order with order number 12345' })
         done()
@@ -523,8 +541,38 @@ describe('actions', () => {
         jest.spyOn(orderService, 'buildFormData').mockImplementation(() => new FormData())
         jest.spyOn(orderService, 'generateOrderNumber').mockImplementation(() => '12345')
         post.mockResolvedValue('success')
-        await actions.save({ state, commit })
+        const response = await actions.save({ state, commit })
+        expect(response).toBe('12345')
         expect(post).toHaveBeenCalledWith('/api/v1/lifelines_order', expect.anything(), true)
+        expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Saved order with order number 12345' })
+        done()
+      })
+    })
+
+    describe('if applicationForm is a fileRef', () => {
+      it('saves order', async (done) => {
+        const commit = jest.fn()
+        const state: ApplicationState = {
+          ...emptyState,
+          order: {
+            orderNumber: '12345',
+            name: null,
+            projectNumber: null,
+            applicationForm: {
+              id: 'id',
+              url: 'url',
+              filename: 'my file'
+            },
+            state: null,
+            submissionDate: null,
+            creationDate: null,
+            updateDate: null
+          }
+        }
+        jest.spyOn(orderService, 'buildFormData').mockImplementation(() => new FormData())
+        post.mockResolvedValue('success')
+        await actions.save({ state, commit })
+        expect(post).toHaveBeenCalledWith('/api/v1/lifelines_order/12345?_method=PUT', expect.anything(), true)
         expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Saved order with order number 12345' })
         done()
       })
@@ -565,6 +613,7 @@ describe('actions', () => {
     describe('if orderNumber is set', () => {
       it('submits the order', async (done) => {
         const commit = jest.fn()
+        const dispatch = jest.fn()
         const state: ApplicationState = {
           ...emptyState,
           order: {
@@ -579,14 +628,17 @@ describe('actions', () => {
           }
         }
         post.mockResolvedValue('success')
-        await actions.submit({ state, commit })
+        await actions.submit({ state, commit, dispatch })
         expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Submitted order with order number 12345' })
+        expect(dispatch).toHaveBeenCalledWith('givePermissionToOrder')
+        expect(dispatch).toHaveBeenCalledWith('sendSubmissionTrigger')
         done()
       })
     })
     describe('if orderNumber not yet set', () => {
       it('submits the order', async (done) => {
         const commit = jest.fn()
+        const dispatch = jest.fn()
         const state: ApplicationState = {
           ...emptyState,
           order: {
@@ -601,8 +653,10 @@ describe('actions', () => {
           }
         }
         post.mockResolvedValue('success')
-        await actions.submit({ state, commit })
+        await actions.submit({ state, commit, dispatch })
         expect(commit).toHaveBeenCalledWith('setToast', { type: 'success', message: 'Submitted order with order number 12345' })
+        expect(dispatch).toHaveBeenCalledWith('givePermissionToOrder')
+        expect(dispatch).toHaveBeenCalledWith('sendSubmissionTrigger')
         done()
       })
     })
@@ -610,9 +664,11 @@ describe('actions', () => {
     describe('when the submission not succesfull', () => {
       let result: any
       let commit: any
+      let dispatch: any
       let state: ApplicationState
       beforeEach(async (done) => {
         commit = jest.fn()
+        dispatch = jest.fn()
         state = {
           ...emptyState,
           order: {
@@ -627,14 +683,61 @@ describe('actions', () => {
           }
         }
         post.mockRejectedValue('error')
-        result = await actions.submit({ commit, state })
+        result = await actions.submit({ commit, state, dispatch })
         done()
       })
 
-      it('should resturn undefined', () => {
+      it('should return undefined', () => {
         expect(result).toBeUndefined()
         expect(commit).not.toHaveBeenCalledWith('setToast', { type: 'success', message: 'Submitted order with order number 12345' })
+        expect(dispatch).not.toHaveBeenCalledWith('givePermissionToOrder')
+        expect(dispatch).not.toHaveBeenCalledWith('sendSubmissionTrigger')
       })
+    })
+  })
+
+  describe('givePermissionToOrder', () => {
+    let state: any
+    beforeEach(async (done) => {
+      state = {
+        order: {
+          orderNumber: '3333'
+        }
+      }
+      await actions.givePermissionToOrder({ state, commit: jest.fn() })
+      done()
+    })
+    it('should resturn undefined', () => {
+      expect(post).toHaveBeenCalledWith('/api/v1/lifelines_order', expect.anything(), true)
+    })
+  })
+
+  describe('sendSubmissionTrigger', () => {
+    let mockPost = jest.fn()
+    beforeEach(async (done) => {
+      mockPost.mockResolvedValue('yep yep')
+      axios.post = mockPost
+      await actions.sendSubmissionTrigger()
+      done()
+    })
+    it('should send a trigger of type submit', () => {
+      expect(mockPost).toHaveBeenCalledWith('/edge-server/trigger?type=submit')
+    })
+  })
+
+  describe('sendSubmissionTrigger error handling', () => {
+    let mockPost = jest.fn()
+    beforeEach(async (done) => {
+      // @ts-ignore
+      global.console = { warn: jest.fn() }
+      mockPost.mockRejectedValue('my warning')
+      axios.post = mockPost
+      await actions.sendSubmissionTrigger()
+      done()
+    })
+    it('should catch and warn, not bother user', () => {
+      expect(console.warn).toBeCalledWith('Send submit trigger failed')
+      expect(console.warn).toBeCalledWith('my warning')
     })
   })
 })