+++ tests/unit/views/CartView.spec.ts
@@ -1,14 +1,26 @@
-import { shallowMount, createLocalVue } from '@vue/test-utils'
-import VueRouter from 'vue-router'
+import { shallowMount, createLocalVue, RouterLinkStub } from '@vue/test-utils'
+
 import CartView from '@/views/CartView.vue'
 import Vuex from 'vuex'
 import Vue from 'vue'
+import flushPromises from 'flush-promises'
+
 Vue.filter('i18n', (value: string) => value) // Add dummy filter for i18n
 
 describe('CartView.vue', () => {
   const localVue = createLocalVue()
   localVue.use(Vuex)
-  localVue.use(VueRouter)
+
+  let stubs = {
+    RouterLink: RouterLinkStub
+  }
+
+  let mocks = {
+    $router: {
+      push: jest.fn()
+    }
+  }
+
   let store: any
   let actions: any
 
@@ -19,7 +31,6 @@ describe('CartView.vue', () => {
     }
 
     state = {
-      isSignedIn: true,
       gridSelection: {
         123: [1, 2],
         456: [3]
@@ -59,15 +70,17 @@ describe('CartView.vue', () => {
   })
 
   it('renders cart view', () => {
-    const wrapper = shallowMount(CartView, { store, localVue })
+    const wrapper = shallowMount(CartView, { stubs, store, localVue })
     expect(wrapper.find('#cart-view').exists()).toBeTruthy()
     expect(wrapper.findAll('li').at(0).text()).toEqual('var 123 ( assessment1, assessment3 )')
     expect(wrapper.findAll('li').at(1).text()).toEqual('var 456 ( assessment3 )')
   })
 
-  it('renders an save button that saves the current state', () => {
-    const wrapper = shallowMount(CartView, { store, localVue })
+  it('renders a save button that saves the current state', async () => {
+    actions.save.mockResolvedValue('12345')
+    const wrapper = shallowMount(CartView, { stubs, store, localVue, mocks })
     wrapper.find('.save').trigger('click')
-    expect(actions.save).toHaveBeenCalled()
+    await flushPromises()
+    expect(mocks.$router.push).toHaveBeenCalledWith({ name: 'load', params: { orderNumber: '12345' } })
   })
 })